+++ e2e-tests/utils/enterFormData.js
@@ -1,10 +1,25 @@
 const { scrollToElement } = require('./scrollToElement');
 
 module.exports = {
-  enterFormData: async function (browser, selectorArgs, input) {
+  enterFormValue: async function (browser, id, input) {
+    const selectorArgs = ['css selector', `#${id}`];
     await scrollToElement(browser, selectorArgs);
     await browser.clearValue(...selectorArgs);
     await browser.setValue(...selectorArgs, input);
-    await browser.pause(5000); // Wait for change... annoying but better than flaky tests
+    await browser.pause(1000); // Wait for change... annoying but better than flaky tests
   },
+  clickFormEl: async function (browser, id, opts) {
+    const { clickLabel, clickChildIndex } = opts || {};
+    const selectorArgs = ['css selector', `#${id}`];
+    if (clickLabel) {
+      selectorArgs[1] = `label[for=${id}]`;
+    }
+    await scrollToElement(browser, selectorArgs);
+    if (clickChildIndex) {
+      selectorArgs[1] = `#${id} *:nth-child(${clickChildIndex})`;
+    }
+    await browser.click(...selectorArgs);
+    await browser.pause(1000); // Wait for change... annoying but better than flaky tests
+  },
+  // TODO: keyboard interactivity
 };
